feat: implement useData hook
